\name{AbPval}
\alias{AbPval}

\title{Determination of the percentile of r and r-squared, by correlation.  Now designated "Abernethy's P-value"}

\description{
Dr. Abernethy has devoted several decades to promoting the percentile of r and r-squared (prr) generated by pivotal Monte Carlo analysis to represent a goodness of fit measure of high merit.  In recognition of this contribution to Weibull Analysis for reliability study, Project "Abernethy Reliability Methods" is attributing Dr. Abernethy's name to this measure.
}

\usage{
AbPval(F,R2,model="weibull")
}

\arguments{
  \item{F}{The quantity of complete failure data points under consideration.}
  \item{R2}{The square of the correlation coefficient derived from residuals of the linear model.}
  \item{model}{A string defining the distribution under consideration. Only a value of "lnorm" will be treated any differently from default of "weibull".}

}

\details{The value returned is derived from a correlation developed from previously run pivotal analysis with 10^8 random samples.  Project "Abernethy Reliability Methods" has judged that only the prr derived from 2 parameter models to have usefullness in comparitive analysis.  This is seen from the "Detect Power" presentations in Appendix D of "The New Weibull Handbook, Fifth Edition".  For validity of a 3rd parameter optimization on a given model over its 2 parameter fit, only the Likelihood Ratio Test will be applied.  
}

\value{
Returns a vector containing the Abernethy P-value and the square of CCC (for comparison with R squared).
}

\references{
  Dr. Robert B. Abernethy, (2008) "The New Weibull Handbook, Fifth Edition"
}

\examples{
AbernethyPvalue<-AbPval(50, 0.996, "lnorm")
}

\keyword{ regression reliability }