\name{lslr}
\alias{lslr}

\title{Least squares linear regression with many optional methods}

\description{
This function implements a wide assortment of options for linear regression fitting of distributions specific to reliability analysis for the Abernethy Reliability Methods project.
}

\usage{
lslr(x, dist="weibull", npar=2, reg_method="xony") 
% We agreed on getting rid of the distinction dist & npar...
}

\arguments{
  \item{x}{A dataframe such as generated by getPPP with column names \code{'data'} and \code{'ppp'}.}
  \item{dist}{A string defining a distribution to be fit. Supported distributions are \code{"weibull"} (default), \code{"lognormal"}, and \code{"gumbel"} (Extreme Value Type 1)}
  \item{npar}{Number of parameters to evaluate. This really is only looking for potential value of 3 for 3rd parameter optimization.  Any other value will resolve the same as default of 2.}
  \item{reg_method}{A string defining a method of regression optimization whether \code{"xony"} axis (default and "best practice" for standard small sample fitting) or \code{"yonx"} axis for alternative study (\code{"yonx"} is recommended for an inspection option by Abernethy).}
}


\value{
Returns a named vector with parameters of the distribution fit and goodness of fit in terms of r squared.
}

\references{
  Dr. Robert B. Abernethy, (2008) \emph{The New Weibull Handbook}, Fifth Edition
  
  Leonard C. Johnson (1964) \emph{The Statistical Treatment of Fatique Experiments}
}

\examples{
fail <-c(90,96,30,49,82)
susp <-c(100,45,10)
message("weibull,   2 parameter:")
print(fit.weibull2p <- lslr(getPPP(fail,susp)))
message("weibull,   3 parameter:")
print(fit.weibull3p <- lslr(getPPP(fail,susp),npar=3))
message("lognormal, 2 parameter:")
print(fit.lognormal2p   <- lslr(getPPP(fail,susp),dist="lognormal"))
message("lognormal, 3 parameter:")
print(fit.lognormal3p   <- lslr(getPPP(fail,susp),dist="lognormal",npar=3))
}

\keyword{ regression reliability }