\name{lslr}
\alias{lslr}

\title{Least squares linear regression with many optional methods}

\description{
This function implements a wide assortment of options for linear regression fitting of distributions specific to reliability analysis for the Abernethy Reliability Methods project.
}

\usage{
lslr(x, dist="weibull", npar=2, reg_method="XonY") 
}

\arguments{
  \item{x}{A dataframe such as generated by getPPP with column names 'data' and 'ppp'.}
  \item{dist}{A string defining a distribution to be fit. Implemented distributions are "weibull" (default), "lnorm", and "gumbel" (Extreme Value Type 1)}
  \item{npar}{Number of parameters to evaluate. This really is only looking for potential value of 3 for 3rd parameter optimization.  Any other value will resolve the same as default of 2.}
  \item{reg_method}{A string defining a method of regression optimization whether "XonY" axis (default and "best practice" for standard small sample fitting) or "YonX" axis for alternative study ("YonX" recommended for an inspection option by Abernethy).}
}


\value{
Returns a named vector with parameters of the distribution fit and goodness of fit in terms of R squared.
}

\references{
  Dr. Robert B. Abernethy, (2008) "The New Weibull Handbook, Fifth Edition"
  Leonard C. Johnson (1964) "The Statistical Treatment of Fatique Experiments"  
}

\examples{
fail <-c(90,96,30,49,82)
susp <-c(100,45,10)
message("weibull,   2 parameter:")
print(fit.weibull2p <- lslr(getPPP(fail,susp)))
message("weibull,   3 parameter:")
print(fit.weibull3p <- lslr(getPPP(fail,susp),npar=3))
message("lognormal, 2 parameter:")
print(fit.lnorm2p   <- lslr(getPPP(fail,susp),dist="lnorm"))
message("lognormal, 3 parameter:")
print(fit.lnorm3p   <- lslr(getPPP(fail,susp),dist="lnorm",npar=3))
}

\keyword{ regression reliability }